move harness to executor
